continued work on issue #83
